+++ client/components/Links/체육회 소개 components/중앙 조직도.tsx
@@ -1,8 +1,8 @@
-import '../../styles/Sub-pages/subpage.css'
-import '../../styles/Sub-pages/중앙조직도-content.css'
-import Dropdownmenu from '../Dropdown/Dropdownmenu'
-import Footer from '../Footer'
-import Header from './Link-header'
+import '../../../styles/Sub-pages/subpage.css'
+import '../../../styles/Sub-pages/체육회 소개 css/중앙조직도-content.css'
+import Dropdownmenu from '../../Dropdown/Dropdownmenu'
+import Footer from '../../Footer'
+import Header from '../Link-header'
 import { MdHome } from 'react-icons/md'
 import { BiSolidRightArrow } from 'react-icons/bi'
 import { Link } from 'react-router-dom'
