+++ client/routes.tsx
@@ -6,9 +6,9 @@ import {
 
 import App from './components/App'
 import { Homepage } from './components/Homepage'
-import { 총재인사말 } from './components/Links/총재인사말'
-import { 설립목적 } from './components/Links/설립목적'
-import { 중앙조직도 } from './components/Links/중앙 조직도'
+import { 총재인사말 } from './components/Links/체육회 소개 components/총재인사말'
+import { 설립목적 } from './components/Links/체육회 소개 components/설립목적'
+import { 중앙조직도 } from './components/Links/체육회 소개 components/중앙 조직도'
 
 export const router = createBrowserRouter(
   createRoutesFromElements(
